move over rtc init to api.ts
